+++ src/libraries/System.IO.Pipelines/src/System/IO/Pipelines/BufferSegment.cs
@@ -8,7 +8,7 @@ using System.Runtime.CompilerServices;
 
 namespace System.IO.Pipelines
 {
-    internal sealed class BufferSegment : ReadOnlySequenceSegment<byte>
+    internal sealed partial class BufferSegment : ReadOnlySequenceSegment<byte>
     {
         private object? _memoryOwner;
         private BufferSegment? _next;
@@ -57,34 +57,6 @@ namespace System.IO.Pipelines
         {
             _memoryOwner = arrayPoolBuffer;
             AvailableMemory = arrayPoolBuffer;
-        }
-
-        public void ResetMemory()
-        {
-            object memoryOwner = _memoryOwner!;
-
-            // Order of below field clears is significant as it clears in a sequential order
-            // https://github.com/dotnet/corefx/pull/35256#issuecomment-462800477
-            Next = null;
-            RunningIndex = 0;
-            Memory = default;
-            _memoryOwner = null;
-            _next = null;
-            _end = 0;
-            AvailableMemory = default;
-
-            // Return the memory, use a fast exact type check rather than checking the inheritance hierarchy
-            // or following the interface mapping.
-            if (memoryOwner.GetType() == typeof(byte[]))
-            {
-                byte[] poolArray = Unsafe.As<byte[]>(memoryOwner);
-                ArrayPool<byte>.Shared.Return(poolArray);
-            }
-            else
-            {
-                Debug.Assert(memoryOwner is IMemoryOwner<byte>);
-                Unsafe.As<IMemoryOwner<byte>>(memoryOwner).Dispose();
-            }
         }
 
         // Exposed for testing
